Finish EDA (#10)
* Fix SetPBPAdvanced

Error in the code where sometimes due to the ifelse statement in the morey variable in the last summarize, would double count a possession, leading to massively inflated numbers.

* Update column names

changed _o to _team, _d to _opp for clarity

* add oRating/dRating, fix consistency with suffixes

* fix game order

wasn't ordering by game date, causing issues with the cumulative stats

* Create GetVenueDistances.R

* Update GetHistoricalData.R

change to only do historical season, only meant to run once

* Create UpdateCurrentSeasonData.R

made to run continuously throughout the season to update the current season's data

* Create BuildInputDF.R

* add a running WinPct

* Update CleaningUtilities.R

Issue where steals, assists, and blocks were spitting NAs. Also found that the NBA changed the wording of the PBP, added additional logic to account for the difference (only affected pre-2006 data)

* changes for new input dataframes

now outputs both centered/scaled (by date) and just a simple lagged dataframe, both to mimic the "snapshot" of the season on any given gameday.

* Fix merging issue

The rows were not aligning correctly because one of the date columns was the wrong datatype. This fixes that issue.

* Wins Flipped

wins were flipped in the logic

* finish EDA